Moving scripts into scripts folder
