--- Assets/Attack.cs
@@ -1,6 +1,0 @@
-public struct Attack
-{
-    public int damageAmount;
-    public Critter target;
-    public Critter sender;
-}
