--- Assets/Attack.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: db514aab57f946529a844c23b56386d2
-timeCreated: 1657239647
