--- Assets/BattleEnvironment.cs
@@ -1,4 +1,0 @@
-public class BattleEnvironment
-{
-
-}
