--- Assets/BattleSceneInitData.cs
@@ -1,5 +1,0 @@
-public static class BattleSceneInitData
-{
-    public static Character[] Characters;
-    public static BattleEnvironment Environment;
-}
