--- Assets/BattleSceneInitData.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 6da22734080c49fe90aa9878dce02e6b
-timeCreated: 1657238863
