--- Assets/ICharacterBattleManager.cs
@@ -1,5 +1,0 @@
- public interface ICharacterBattleManager
-{
-    public void AttemptFlee();
-}
-
