+++ Assets/Scripts/Attack.cs
@@ -1,0 +1,6 @@
+public struct Attack
+{
+    public int damageAmount;
+    public Critter target;
+    public Critter sender;
+}
