+++ Assets/Scripts/BattleEnvironment.cs
@@ -1,0 +1,4 @@
+public class BattleEnvironment
+{
+
+}
