+++ Assets/Scripts/BattleSceneInitData.cs
@@ -1,0 +1,5 @@
+public static class BattleSceneInitData
+{
+    public static Character[] Characters;
+    public static BattleEnvironment Environment;
+}
