+++ Assets/Scripts/ICharacterBattleManager.cs
@@ -1,0 +1,5 @@
+ public interface ICharacterBattleManager
+{
+    public void AttemptFlee();
+}
+
