+++ Assets/Scripts/ICritterBattleManager.cs
