+++ Assets/Scripts/ICritterBattleManager.cs.meta
